# tpiaew/swagger/AlquilerVehiculos.yaml
---
swagger: "2.0"
info:
  title: "AlquilerVehiculosREST"
  description: "Documentación del servicio AlquilerVehiculos desarrollado por el grupo 3, conformado por Fascioli, Bargiano, Ponce, Micheletti."
  version: "1.0"
basePath: "/"
schemes:
- "http"
paths:
  /paises:
    get:
      description: "Obtener paises."
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              paises:
                type: "array"
                items:
                  $ref: "#/definitions/Pais"
  /ciudades:
    get:
      description: "Obtener ciudades."
      parameters:
      - name: "idPais"
        in: "query"
        type: "integer"
        format: "int32"
        required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              ciudades:
                type: "array"
                items:
                  $ref: "#/definitions/Ciudad"
  /vehiculosDisponibles:
    get:
      description: "Obtener los vehiculos disponibles para una ciudad y en un periodo de tiempo dado."
      parameters:
      - name: "idCiudad"
        in: "query"
        type: "integer"
        format: "int32"
        required: true
      - name: "fechaRetiro"
        in: "query"
        type: "string"
        required: true
      - name: "fechaDevolucion"
        in: "query"
        type: "string"
        required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              vehiculos:
                type: "array"
                items:
                  $ref: "#/definitions/Vehiculo"
  /reserva:
    post:
      description: "Realizar la reserva de un vehiculo."
      parameters:
      - in: "body"
        name: "payload"
        schema:
          type: "object"
          properties:
            apellidoNombre:
              type: "string"
            fechaDevolucion:
              type: "string"
            fechaRetiro:
              type: "string"
            idVehiculoCiudad:
              type: "number"
            lugarDevolucionRetiro:
              type: "string"
            lugarRetiro:
              type: "string"
            nroDocumento:
              type: "string"
      responses:
        201:
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              reservamongo:
                $ref: "#/definitions/ReservaMongo"
  /reserva/lista:
    get:
      description: "Obtener las reservas de un cliente."
      parameters:
      - name: "idCliente"
        in: "query"
        type: "integer"
        format: "int64"
        required: true
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              vehiculos:
                type: "array"
                items:
                  $ref: "#/definitions/ReservaMongo"
  /reserva/baja:
    post:
      description: "Cancela una reserva de un vehiculo."
      parameters:
      - in: "body"
        name: "payload"
        schema:
          type: "object"
          properties:
            codigoReserva:
              type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
definitions:
    Pais:
      type: "object"
      properties:
        b:Id:
          type: "string"
        b:Nombre:
          type: "string"
    Ciudad:
      type: "object"
      properties:
        b:Id:
          type: "string"
        b:Nombre:
          type: "string"
        b:PaisEntity:
          $ref: "#/definitions/Pais"
    ReservaMongo:
      type: "object"
      properties:
        codigoReserva:
          type: "string"
        fechaReserva:
          type: "string"
          format: "date-time"
        idCliente:
          type: "integer"
          format: int64
        costo:
          type: "integer"
          format: int32
        precioVenta:
          type: "integer"
          format: int32
    Vehiculo:
      type: "object"
      properties:
        a:CantidadDisponible:
          type: "number"
        a:CantidadPuertas:
          type: "number"
        a:CiudadId:
          type: "number"
        a:Id:
          type: "number"
        a:Marca:
          type: "string"
        a:Modelo:
          type: "string"
        a:PrecioPorDia:
          type: "string"
        a:Puntaje:
          type: "string"
        a:TieneAireAcon:
          type: "boolean"
        a:TieneDireccion:
          type: "boolean"
        a:TipoCambio:
          type: "string"
        a:VehiculoCiudadId:
          type: "number"
        precioDeVenta:
          type: "number"